+++ src/main/java/com/procedures/demo/customer/CustomerRepository.java
@@ -2,4 +2,4 @@ package com.procedures.demo.customer;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface CustomerRepository extends CrudRepository<Customer, Integer> {}
+public interface CustomerRepository extends CrudRepository<Customer, Long> {}
